📑 :: 업적달성 엔티티 수정
